+++ .changeset/afraid-masks-dress.md
@@ -1,0 +1,5 @@
+---
+'@keystone-next/keystone': major
+---
+
+Moved exports of `@keystone-next/keystone` to `@keystone-next/keystone/system`
