+++ packages/keystone/package.json
@@ -19,7 +19,8 @@
     "admin-ui",
     "testing",
     "fields",
-    "types"
+    "types",
+    "system"
   ],
   "bin": {
     "keystone-next": "bin/cli.js"
@@ -136,7 +137,7 @@
   },
   "preconstruct": {
     "entrypoints": [
-      "index.ts",
+      "system.ts",
       "next.ts",
       "___internal-do-not-use-will-break-in-patch/{node-api,next-graphql}.ts",
       "___internal-do-not-use-will-break-in-patch/admin-ui/pages/*/index.tsx",
